+++ rtl/audio/mp2_pcm_fifo.sv
@@ -3,7 +3,8 @@ module mp2_pcm_fifo (
     input wire [66:0] wr_data,
     input wire wr_en,rd_en,
     output wire wr_full,rd_empty,
-    output wire [66:0] rd_data
+    output wire [66:0] rd_data,
+    output wire [11:0] wr_used
 );
 dcfifo #(
     .lpm_numwords(4096),.lpm_showahead("ON"),.lpm_type("dcfifo"),
@@ -11,5 +12,5 @@ dcfifo #(
     .underflow_checking("ON"),.use_eab("ON"),.rdsync_delaypipe(4),
     .wrsync_delaypipe(4),.write_aclr_synch("ON"),.read_aclr_synch("ON")
 ) fifo (.aclr(reset),.data(wr_data),.wrclk(wr_clk),.wrreq(wr_en),
-    .wrfull(wr_full),.q(rd_data),.rdclk(rd_clk),.rdreq(rd_en),.rdempty(rd_empty));
+    .wrfull(wr_full),.wrusedw(wr_used),.q(rd_data),.rdclk(rd_clk),.rdreq(rd_en),.rdempty(rd_empty));
 endmodule